+++ src/main/java/net/chesstango/piazzolla/syzygy/PieceSymmetricWdl.java
@@ -28,7 +28,7 @@ class PieceSymmetricWdl extends TableBase {
         // Next, there may be a padding byte to align the position within the tablebase file to a multiple of 2 bytes.
         data.ptr += data.ptr & 1;
 
-        int[] size_white = new int[3];
+        long[] size_white = new long[3];
 
         ei_wtm.precomp = new PairsData(WDL, data, tb_size_white, size_white);
